Add parsing of action files
